Refactor to follow DRY principle - centralize StorageLocationType enum
Moved StorageLocationType enum to a shared constants.py module to eliminate
duplication and prevent future inconsistencies. The service layer now derives
VALID_TYPES from the enum instead of maintaining a separate hardcoded list.

Changes:
- Create backend/src/constants.py with StorageLocationType enum
- Update backend/src/api/storage.py to import from constants
- Update backend/src/services/storage_service.py to import from constants
- Service VALID_TYPES now derived from enum: {t.value for t in StorageLocationType}
- Maintains single source of truth for valid location types
