+++ backend/src/constants.py
@@ -1,0 +1,18 @@
+"""
+Shared constants and enums for the application.
+"""
+
+from enum import Enum
+
+
+class StorageLocationType(str, Enum):
+    """Valid storage location types matching database constraints."""
+
+    CONTAINER = "container"
+    ROOM = "room"
+    BUILDING = "building"
+    CABINET = "cabinet"
+    DRAWER = "drawer"
+    SHELF = "shelf"
+    BIN = "bin"
+    BAG = "bag"
